ImageNormalization2D Layer replaces process_image
ImageNormalization2D has been used in each model in the model zoo. This will be an issue for 3D data. filter_size is passed as variable to models if it is not pre-defined in the architecture.

ImageNormalization2D takes norm_method and filter_size as params, where norm_method is 'std', 'max', or 'median'.  Median normalization needs to be confirmed, as it uses tf.contrib function.