pulled TrackingGeneratorTests notebook changes
modified training, model_zoo, image_generators, and tracking to
reflect the changes in the TrackingGeneratorTests notebook.

changes were tested by training a new model.

** image_generators now imports keras_preprocessing **
In the newest version of tensorflow we can remove this import as
tensorflow 1.11+ depends on keras_preprocessing.

removed trailing whitespace

tracking.cell_tracker now has a features argument to take in
the same set of features as the siamese_model

small updates to the tracking notebooks to make sure everything
works properly.